ILT-812

Approved-by: maksim.zakharov <mz@smartvalley.io>
Approved-by: Valentin Livkin <vl@icolab.com>

+++ SmartValley.WebApi/Estimates/Responses/EstimateResponse.cs
@@ -12,7 +12,7 @@ namespace SmartValley.WebApi.Estimates.Responses
         {
             return new EstimateResponse
                    {
-                       Score = estimate.Score,
+                       Score = (int) estimate.Score,
                        Comment = estimate.Comment
                    };
         }
